+++ src/components/formBuilder/FormBuilder.test.tsx
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { getExampleBlockConfig } from "@/pageEditor/tabs/editTab/exampleBlockConfigs";
+import { getExampleBlockConfig } from "@/pageEditor/exampleBlockConfigs";
 import {
   createFormikTemplate,
   fireTextInput,
