+++ src/pageEditor/exampleBlockConfigs.ts
